refactor modal classes, relating reviews, businesses and users to each other  [Fixes #155327560]
